requirements.txt
streamlit
python-docx
PyPDF2
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters
langchain-ollama
chromadb
python-dotenv
opencv-python
groq
Pyrebase4==4.7.1
setuptools==70.0.0
pycryptodome
firebase-admin
pytesseract
pdf2image
Pillow
numpy
scikit-learn
requests
google-search-results
sentence-transformers
huggingface-hub
rank_bm25
